# lib/api-zod/src/generated/types/chatResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatResultSource } from './chatResultSource';
import type { ChatResultStatus } from './chatResultStatus';

export interface ChatResult {
  status: ChatResultStatus;
  message: string;
  model?: string;
  fallback: boolean;
  source: ChatResultSource;
}