# lib/api-zod/src/generated/types/ecsStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EcsStatusMavenProject } from './ecsStatusMavenProject';
import type { EcsStatusProcess } from './ecsStatusProcess';
import type { EcsStatusStatus } from './ecsStatusStatus';

export interface EcsStatus {
  status: EcsStatusStatus;
  mavenProject: EcsStatusMavenProject;
  process: EcsStatusProcess;
}