# lib/api-zod/src/generated/types/fileContentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FileContentInput {
  /**
     * @minLength 1
     * @maxLength 512
     */
  path: string;
  /** @maxLength 1048576 */
  content: string;
}