# lib/api-zod/src/generated/types/receiptEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReceiptEventMetadata } from './receiptEventMetadata';

export interface ReceiptEvent {
  id: string;
  type: string;
  createdAt: Date;
  metadata?: ReceiptEventMetadata;
}