# lib/api-zod/src/generated/types/systemStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SystemStatusDatabase } from './systemStatusDatabase';
import type { SystemStatusProviders } from './systemStatusProviders';
import type { SystemStatusStatus } from './systemStatusStatus';

export interface SystemStatus {
  status: SystemStatusStatus;
  database: SystemStatusDatabase;
  providers: SystemStatusProviders;
}