# lib/api-zod/src/generated/types/workingFile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkingFileParseStatus } from './workingFileParseStatus';

export interface WorkingFile {
  id: string;
  name: string;
  size: number;
  contentType: string;
  createdAt: Date;
  parseStatus: WorkingFileParseStatus;
  /** Bounded text derived from the original bytes; never the original file. */
  parsedContent?: string;
}