# lib/api-zod/src/generated/types/workingFileUploadRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkingFileUploadRequest {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 10485760
     */
  size: number;
  /**
     * @minLength 1
     * @maxLength 127
     */
  contentType: string;
}