# lib/api-zod/src/generated/types/workingFileUploadUrl.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkingFileUploadUrl {
  uploadURL: string;
  objectPath: string;
  expiresInSeconds: number;
}