# lib/api-zod/src/generated/types/workspaceFile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkspaceFile {
  path: string;
  /**
     * @minimum 0
     * @maximum 1048576
     */
  size: number;
  modifiedAt: Date;
}